moved firebase utils in seperate folder
